requirements.txt
Flask==3.0.0
msal==1.25.0
python-dotenv==1.0.0
